+++ pyminflux/ui/importer.py
@@ -14,7 +14,6 @@
 #
 from typing import Sequence
 
-import numpy as np
 from PySide6.QtCore import Slot
 from PySide6.QtWidgets import QDialog
 
@@ -32,7 +31,7 @@ class Importer(QDialog, Ui_Importer):
         relocalizations: Sequence,
         dwell_time: float,
         is_tracking: bool,
-        bin_dcr: bool,
+        pool_dcr: bool,
     ):
         # Call the base class
         super().__init__()
@@ -50,7 +49,7 @@ class Importer(QDialog, Ui_Importer):
         self._relocalizations = relocalizations
         self._dwell_time = dwell_time
         self._is_tracking = is_tracking
-        self._bin_dcr = bin_dcr
+        self._pool_dcr = pool_dcr
 
         # Set the dwell time
         self.ui.leDwellTime.setText(f"{self._dwell_time}")
@@ -59,7 +58,7 @@ class Importer(QDialog, Ui_Importer):
         self.ui.cbTracking.setChecked(self._is_tracking)
 
         # Set the bin DCR checkbox
-        self.ui.cbBinDCR.setChecked(self._bin_dcr)
+        self.ui.cbBinDCR.setChecked(self._pool_dcr)
 
         # Characters
         self.valid_char = "✓"
@@ -119,7 +118,7 @@ class Importer(QDialog, Ui_Importer):
     def set_connections(self):
         self.ui.leDwellTime.textChanged.connect(self.persist_dwell_time)
         self.ui.cbTracking.stateChanged.connect(self.persist_is_tracking)
-        self.ui.cbBinDCR.stateChanged.connect(self.persist_bin_dcr)
+        self.ui.cbBinDCR.stateChanged.connect(self.persist_pool_dcr)
         self.ui.pb_last_valid.clicked.connect(self.set_last_valid)
         for i in range(len(self.widgets_list)):
             self.widgets_list[i][0].setProperty("index", i)
@@ -138,8 +137,8 @@ class Importer(QDialog, Ui_Importer):
         self._is_tracking = state != 0
 
     @Slot(int)
-    def persist_bin_dcr(self, state):
-        self._bin_dcr = state != 0
+    def persist_pool_dcr(self, state):
+        self._pool_dcr = state != 0
 
     @Slot()
     def set_last_valid(self):
@@ -191,7 +190,7 @@ class Importer(QDialog, Ui_Importer):
             "cfr_iteration": self._cfr_iteration,
             "is_tracking": self._is_tracking,
             "dwell_time": self._dwell_time,
-            "bin_dcr": self._bin_dcr,
+            "pool_dcr": self._pool_dcr,
         }
 
     def hide_to(self, to_value: int):